requirements.txt
Flask>=2.0
PyPDF2>=3.0
joblib
requests
google-generativeai            # opcional, se usar Gemini
openai>=0.27.0                 # opcional, se usar OpenAI
scikit-learn                   # se for treinar/usar local
